Move RSVP-related settings to new RSVP settings panel and online event link to Venue settings for better organization.
