+++ src/panels/event-settings/index.js
@@ -15,12 +15,8 @@ import { PluginDocumentSettingPanel } from '@wordpress/editor';
  * Internal dependencies.
  */
 import { isEventPostType } from '../../helpers/event';
-import AnonymousRsvpPanel from './anonymous-rsvp';
 import DateTimeRangePanel from './datetime-range';
-import GuestLimitPanel from './guest-limit';
-import MaxAttendanceLimitPanel from './max-attendance-limit';
 import NotifyMembersPanel from './notify-members';
-import OnlineEventLinkPanel from './online-link';
 import { EventPluginDocumentSettings } from './slot';
 
 /**
@@ -48,10 +44,6 @@ const EventSettings = () => {
 
 				<VStack spacing={ 4 }>
 					<DateTimeRangePanel />
-					<OnlineEventLinkPanel />
-					<GuestLimitPanel />
-					<MaxAttendanceLimitPanel />
-					<AnonymousRsvpPanel />
 					<NotifyMembersPanel />
 				</VStack>
 			</PluginDocumentSettingPanel>
