--- src/panels/event-settings/max-attendance-limit/index.js
@@ -1,14 +1,0 @@
-/**
- * Internal dependencies.
- */
-import MaxAttendanceLimit from '../../../components/MaxAttendanceLimit';
-
-const MaxAttendanceLimitPanel = () => {
-	return (
-		<section>
-			<MaxAttendanceLimit />
-		</section>
-	);
-};
-
-export default MaxAttendanceLimitPanel;
