+++ src/panels/index.js
@@ -2,4 +2,5 @@
  * Internal dependencies.
  */
 import './event-settings';
+import './rsvp-settings';
 import './venue-settings';
