+++ src/panels/venue-settings/fill.js
@@ -7,6 +7,10 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
+import {
+	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
+	__experimentalVStack as VStack,
+} from '@wordpress/components';
 import { PluginDocumentSettingPanel } from '@wordpress/editor';
 
 /**
@@ -14,6 +18,7 @@ import { PluginDocumentSettingPanel } from '@wordpress/editor';
  */
 import { VenuePluginDocumentSettings } from './slot';
 import { isEventPostType } from '../../helpers/event';
+import OnlineEventLinkPanel from './online-link';
 
 export default function VenuePluginFill() {
 	return (
@@ -24,6 +29,10 @@ export default function VenuePluginFill() {
 				className="gatherpress-venue-settings"
 			>
 				<VenuePluginDocumentSettings.Slot />
+
+				<VStack spacing={ 4 }>
+					<OnlineEventLinkPanel />
+				</VStack>
 			</PluginDocumentSettingPanel>
 		)
 	);
