+++ src/panels/venue-settings/index.js
@@ -2,8 +2,6 @@
  * WordPress dependencies.
  */
 import { __ } from '@wordpress/i18n';
-import domReady from '@wordpress/dom-ready';
-import { dispatch, select } from '@wordpress/data';
 import {
 	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
 	__experimentalVStack as VStack,
@@ -63,35 +61,4 @@ registerPlugin( 'gatherpress-venue-settings', {
 
 registerPlugin( 'gatherpress-venue-settings-at-events', {
 	render: VenuePluginFill,
-} );
-
-/**
- * Toggle Venue Settings Panel
- *
- * This script ensures that the venue settings panel is open in the WordPress block editor.
- * It uses the `domReady` function to ensure the DOM is ready before execution.
- * If the venue settings panel is not open, it opens the venue settings panel using
- * the `toggleEditorPanelOpened` function.
- *
- * @since 1.0.0
- *
- * @return {void}
- */
-domReady( () => {
-	const selectEditPost = select( 'core/edit-post' );
-	const dispatchEditor = dispatch( 'core/editor' );
-
-	if ( ! selectEditPost || ! dispatchEditor ) {
-		return;
-	}
-
-	const isVenuePanelOpened = selectEditPost.isEditorPanelOpened(
-		'gatherpress-venue-settings/gatherpress-venue-settings',
-	);
-
-	if ( ! isVenuePanelOpened ) {
-		dispatchEditor.toggleEditorPanelOpened(
-			'gatherpress-venue-settings/gatherpress-venue-settings',
-		);
-	}
 } );
